Added PATCH in all classes

+++ src/main/java/com/example/restaurantreservationaa/domain/MenuItem.java
@@ -23,7 +23,7 @@ public class MenuItem {
     @Column
     private String description;
     @Column
-    private double price;
+    private Float price;
     @Column
     private String category;
     @Column
